Users, Auth, Login / Logout
Add Accounts context and Users schema, with hashed passwords and
changesets for updating them vs non-sensitive data. Also add Auth plug,
so current_user is avail in conn.assigns if logged in. Add Logout or Login
and Register links to header if user is or is not logged in respectively.