First Contact, a tale of hack and slash
Now have ability to load an entire journal for a user including all
entries (past and upcoming) with prompts and any reflections if they
have made some.

Includes seed data for Li-Anne's course "Mindfulness for Inner Peace".

Beware the nasty hack in session_controller that is hardcoded to
redirect all non-admin users to Li-Anne's a course. It just looks
up the course by name and redirects to /reflect/journal/:id.